split item from view
